requirements.txt
# Core trading libraries
alpaca-py==0.21.0
alpha-vantage==2.3.1

# Data processing and analysis
pandas==2.2.0
numpy==1.26.3
ta==0.11.0

# GCP Cloud Functions
functions-framework==3.5.0

# HTTP requests and API
requests==2.31.0

# Configuration and environment
python-dotenv==1.0.1

# Testing and development
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
